requirements.txt
telethon>=1.27.0
pyyaml>=6.0
